cleaned up query files and moved all helper functions into utils folder
